+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 -   Upgrade node version from 16.13.0 to 18.16.0 (#187)
+-   New icon for list ordering in tables (#191)
 
 ### Fixed